Final project update
