+++ 09-room-db/projects/final/app/src/main/java/com/kodeco/recipefinder/network/SpoonacularService.kt
@@ -1,15 +1,50 @@
+/*
+ * Copyright (c) 2023 Kodeco Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * Notwithstanding the foregoing, you may not use, copy, modify, merge, publish,
+ * distribute, sublicense, create a derivative work, and/or sell copies of the
+ * Software in any work that is designed, intended, or marketed for pedagogical or
+ * instructional purposes related to programming, coding, application development,
+ * or information technology.  Permission for such use, copying, modification,
+ * merger, publication, distribution, sublicensing, creation of derivative works,
+ * or sale is expressly withheld.
+ *
+ * This project and source code may use libraries or frameworks that are
+ * released under various Open-Source licenses. Use of those libraries and
+ * frameworks are governed by their own individual licenses.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.kodeco.recipefinder.network
 
+import com.kodeco.recipefinder.data.models.RecipeInformationResponse
+import com.kodeco.recipefinder.data.models.SearchRecipesResponse
 import com.kodeco.recipefinder.viewmodels.PAGE_SIZE
 import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
 
-const val apiKey = "<Your Key Here>"
+const val apiKey = "<Replace with API Key>"
 
 interface SpoonacularService {
   @GET("recipes/complexSearch?&apiKey=$apiKey")
@@ -17,10 +52,10 @@ interface SpoonacularService {
     @Query("query") query: String,
     @Query("offset") offset: Int,
     @Query("number") number: Int = PAGE_SIZE
-  ): RecipeResponse
+  ): SearchRecipesResponse
 
   @GET("recipes/{id}/information?includeNutrition=false&apiKey=$apiKey")
-  suspend fun queryRecipe(@Path("id") id: Int): SpoonacularRecipe
+  suspend fun queryRecipe(@Path("id") id: Int): RecipeInformationResponse
 }
 
 object RetrofitInstance {
@@ -29,7 +64,6 @@ object RetrofitInstance {
   private fun provideMoshi(): Moshi =
     Moshi
       .Builder()
-      .add(KotlinJsonAdapterFactory())
       .build()
 
   private val retrofit: Retrofit by lazy {
