+++ 09-room-db/projects/final/app/src/main/java/com/kodeco/recipefinder/ui/RecipeDetails.kt
@@ -1,3 +1,37 @@
+/*
+ * Copyright (c) 2023 Kodeco Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * Notwithstanding the foregoing, you may not use, copy, modify, merge, publish,
+ * distribute, sublicense, create a derivative work, and/or sell copies of the
+ * Software in any work that is designed, intended, or marketed for pedagogical or
+ * instructional purposes related to programming, coding, application development,
+ * or information technology.  Permission for such use, copying, modification,
+ * merger, publication, distribution, sublicensing, creation of derivative works,
+ * or sale is expressly withheld.
+ *
+ * This project and source code may use libraries or frameworks that are
+ * released under various Open-Source licenses. Use of those libraries and
+ * frameworks are governed by their own individual licenses.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.kodeco.recipefinder.ui
 
 import android.widget.TextView
@@ -39,7 +73,7 @@ import com.kodeco.recipefinder.LocalNavigatorProvider
 import com.kodeco.recipefinder.LocalPrefsProvider
 import com.kodeco.recipefinder.LocalRepositoryProvider
 import com.kodeco.recipefinder.R
-import com.kodeco.recipefinder.network.SpoonacularRecipe
+import com.kodeco.recipefinder.data.models.RecipeInformationResponse
 import com.kodeco.recipefinder.ui.theme.HeadlineSmall
 import com.kodeco.recipefinder.ui.theme.lighterBlue
 import com.kodeco.recipefinder.ui.widgets.buildRecipeImageBuilder
@@ -55,7 +89,7 @@ fun RecipeDetails(recipeId: Int? = null, databaseRecipeId: Int? = null) {
   val viewModel: RecipeViewModel = viewModel(factory = viewModelFactory {
     RecipeViewModel(prefs)
   })
-  val recipeState = remember { mutableStateOf<SpoonacularRecipe?>(null) }
+  val recipeState = remember { mutableStateOf<RecipeInformationResponse?>(null) }
   val navController = LocalNavigatorProvider.current
   val repository = LocalRepositoryProvider.current
   if (recipeId != null) {
@@ -120,7 +154,7 @@ fun TitleRow(
   modifier: Modifier = Modifier,
   navController: NavHostController,
   viewModel: RecipeViewModel,
-  recipe: SpoonacularRecipe,
+  recipe: RecipeInformationResponse,
   isBookmark: Boolean
 ) {
   val repository = LocalRepositoryProvider.current
