+++ 09-room-db/projects/final/app/src/main/java/com/kodeco/recipefinder/viewmodels/RecipeViewModel.kt
@@ -1,3 +1,37 @@
+/*
+ * Copyright (c) 2023 Kodeco Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * Notwithstanding the foregoing, you may not use, copy, modify, merge, publish,
+ * distribute, sublicense, create a derivative work, and/or sell copies of the
+ * Software in any work that is designed, intended, or marketed for pedagogical or
+ * instructional purposes related to programming, coding, application development,
+ * or information technology.  Permission for such use, copying, modification,
+ * merger, publication, distribution, sublicensing, creation of derivative works,
+ * or sale is expressly withheld.
+ *
+ * This project and source code may use libraries or frameworks that are
+ * released under various Open-Source licenses. Use of those libraries and
+ * frameworks are governed by their own individual licenses.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 package com.kodeco.recipefinder.viewmodels
 
 import androidx.lifecycle.ViewModel
@@ -12,9 +46,9 @@ import com.kodeco.recipefinder.data.recipeDbsToRecipes
 import com.kodeco.recipefinder.data.recipeToDb
 import com.kodeco.recipefinder.data.recipeToSpoonacularRecipe
 import com.kodeco.recipefinder.data.spoonacularIngredientsToIngredients
+import com.kodeco.recipefinder.data.models.RecipeInformationResponse
 import com.kodeco.recipefinder.data.spoonacularRecipeToRecipe
 import com.kodeco.recipefinder.network.RetrofitInstance
-import com.kodeco.recipefinder.network.SpoonacularRecipe
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -48,7 +82,7 @@ class RecipeViewModel(private val prefs: Prefs) : ViewModel() {
   private val _recipeListState = MutableStateFlow<List<Recipe>>(listOf())
   val recipeListState = _recipeListState.asStateFlow()
 
-  private val _recipeState = MutableStateFlow<SpoonacularRecipe?>(null)
+  private val _recipeState = MutableStateFlow<RecipeInformationResponse?>(null)
   val recipeState = _recipeState.asStateFlow()
 
   private val _queryState = MutableStateFlow(QueryState())
@@ -137,7 +171,7 @@ class RecipeViewModel(private val prefs: Prefs) : ViewModel() {
     }
   }
 
-  suspend fun bookmarkRecipe(repository: Repository, recipe: SpoonacularRecipe) {
+  suspend fun bookmarkRecipe(repository: Repository, recipe: RecipeInformationResponse) {
     withContext(Dispatchers.IO) {
       repository.insertRecipe(spoonacularRecipeToRecipe(recipe))
       repository.insertIngredients(
